+++ src/test/java/com/qa/cinema/EndPointTests/ShowingEndPointTest.java
@@ -1,4 +1,4 @@
-package com.qa.cinema.entityTests;
+package com.qa.cinema.EndPointTests;
 
 import org.junit.Assert;
 import com.qa.cinema.rest.ShowingEndPoint;
